+++ api/src/main/java/de/maxhenkel/voicechat/api/audiochannel/EntityAudioChannel.java
@@ -5,14 +5,14 @@ import de.maxhenkel.voicechat.api.Entity;
 public interface EntityAudioChannel extends AudioChannel {
 
     /**
-     * @param whispering if the entity should whisper
-     */
-    void setWhispering(boolean whispering);
-
-    /**
      * @return if the entity is whispering
      */
     boolean isWhispering();
+
+    /**
+     * @param whispering if the entity should whisper
+     */
+    void setWhispering(boolean whispering);
 
     /**
      * Sets a new entity where this channel is attached to.
@@ -25,5 +25,15 @@ public interface EntityAudioChannel extends AudioChannel {
      * @return the entity where the channel is attached to
      */
     Entity getEntity();
+
+    /**
+     * @return the distance, the audio can be heard
+     */
+    float getDistance();
+
+    /**
+     * @param distance the distance, the audio can be heard
+     */
+    void setDistance(float distance);
 
 }