+++ api/src/main/java/de/maxhenkel/voicechat/api/audiochannel/LocationalAudioChannel.java
@@ -16,4 +16,14 @@ public interface LocationalAudioChannel extends AudioChannel {
      */
     Position getLocation();
 
+    /**
+     * @return the distance, the audio can be heard
+     */
+    float getDistance();
+
+    /**
+     * @param distance the distance, the audio can be heard
+     */
+    void setDistance(float distance);
+
 }