+++ common/src/main/java/de/maxhenkel/voicechat/plugins/impl/packets/SoundPacketImpl.java
@@ -9,6 +9,7 @@ import de.maxhenkel.voicechat.plugins.impl.PositionImpl;
 import de.maxhenkel.voicechat.voice.common.GroupSoundPacket;
 import de.maxhenkel.voicechat.voice.common.LocationSoundPacket;
 import de.maxhenkel.voicechat.voice.common.PlayerSoundPacket;
+import de.maxhenkel.voicechat.voice.common.Utils;
 
 import java.util.UUID;
 
@@ -41,16 +42,35 @@ public class SoundPacketImpl implements SoundPacket {
 
     @Override
     public EntitySoundPacket toEntitySoundPacket(UUID entityUuid, boolean whispering) {
-        return new EntitySoundPacketImpl(new PlayerSoundPacket(packet.getSender(), packet.getData(), packet.getSequenceNumber(), whispering));
+        return toEntitySoundPacket(entityUuid, whispering, getDistance());
+    }
+
+    @Override
+    public EntitySoundPacket toEntitySoundPacket(UUID entityUuid, boolean whispering, float distance) {
+        return new EntitySoundPacketImpl(new PlayerSoundPacket(packet.getSender(), packet.getData(), packet.getSequenceNumber(), whispering, distance));
     }
 
     @Override
     public LocationalSoundPacket toLocationalSoundPacket(Position position) {
+        return toLocationalSoundPacket(position, getDistance());
+    }
+
+    @Override
+    public LocationalSoundPacket toLocationalSoundPacket(Position position, float distance) {
         if (position instanceof PositionImpl p) {
-            return new LocationalSoundPacketImpl(new LocationSoundPacket(packet.getSender(), p.getPosition(), packet.getData(), packet.getSequenceNumber()));
+            return new LocationalSoundPacketImpl(new LocationSoundPacket(packet.getSender(), p.getPosition(), packet.getData(), packet.getSequenceNumber(), distance));
         } else {
             throw new IllegalArgumentException("position is not an instance of PositionImpl");
         }
+    }
+
+    private float getDistance() {
+        if (this instanceof EntitySoundPacket p) {
+            return p.getDistance();
+        } else if (this instanceof LocationalSoundPacket p) {
+            return p.getDistance();
+        }
+        return Utils.getDefaultDistance();
     }
 
     @Override